docs/pyproject.toml
[tool.poetry]
name = "scylla"
description = "Scylla Documentation"
version = "4.3.0"
authors = ["ScyllaDB Contributors"]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "6.0.1"
pygments = "2.15.1"
recommonmark = "0.7.1"
sphinx-scylladb-theme = "~1.6.1"
sphinx-sitemap = "2.5.1"
sphinx-autobuild = "2021.3.14"
Sphinx = "7.2.6"
sphinx-multiversion-scylla = "~0.3.1"
sphinx-markdown-tables = "0.0.15"
redirects_cli ="~0.1.2"
sphinxcontrib-datatemplates = "^0.9.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
